Fix 6_Figures.ipynb to run end-to-end (Colab setup cell + QUICK_RUN path)
- Standard setup cell, artifact-discovery globs for trained models,
  data bootstrap reusing cached SHO/loop fits
- Notebook-local violin cells: drop wide-format concat producing
  duplicate-index DataFrames (modern seaborn rejects them), use
  ignore_index=True — same fix as be/viz.py
- viz.py: return fig from three plot methods so composite figure
  cells can reuse them
- Model.predict: cast input batches to float32 (model weights are
  float32; hysteresis data arrives float64)
- Verified headless: nbconvert exit 0, zero error outputs, all 27
  paper figures rendered locally (54 png+svg files)

+++ m3_learning/src/m3_learning/nn/Fitter1D/Fitter1D.py
@@ -531,8 +531,9 @@ class Model(nn.Module):
             if i == num_batches - 1:
                 end = num_elements
 
+            # model weights are float32; input data may arrive as float64
             pred_batch, params_scaled_, params_ = self.model(
-                train_batch.to(self.device))
+                train_batch.to(self.device).float())
 
             if is_SHO:
                 predictions[start:end] = pred_batch.cpu().detach()
